Update fit.py
One of the previous request that we already discussed with Maxime. But got lost in the pool of requests...

Modified the save_result  and the fit_video functions to save more results of the fits (x, y, z, uncertainties, chi2, etc.). 

In save_result: added variables to save.
In fit_video: line 289, shape of the saving, to match save_result